renamed ConvInpGen_1D_custom to ConvInpGen_1D_dws_lowbuffer
